# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Automating resume and cover letter generation, allowing users to easily tailor their applications to specific job opportunities while highlighting their unique skills and experiences.",
        "title": "Next Gen Job Hunting API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/joburl": {
            "get": {
                "description": "Get all job URLs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "joburl"
                ],
                "summary": "Get all job URLs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/joburl.JobUrl"
                            }
                        }
                    },
                    "500": {
                        "description": "Could not fetch job URLs",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new job URL for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "joburl"
                ],
                "summary": "Create a new job URL",
                "parameters": [
                    {
                        "description": "Job URL",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/joburl.JobUrl"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/joburl.JobUrl"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    },
                    "500": {
                        "description": "Could not create Job URL",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    }
                }
            }
        },
        "/joburl/{id}": {
            "get": {
                "description": "Get a job URL by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "joburl"
                ],
                "summary": "Get a job URL by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Job URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/joburl.JobUrl"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    },
                    "404": {
                        "description": "Job URL not found",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    },
                    "500": {
                        "description": "Could not fetch Job URL",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a job URL by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "joburl"
                ],
                "summary": "Update a job URL by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Job URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Job URL",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/joburl.JobUrl"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/joburl.JobUrl"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    },
                    "500": {
                        "description": "Could not update Job URL",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a job URL by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "joburl"
                ],
                "summary": "Delete a job URL by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Job URL ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Job URL deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    },
                    "404": {
                        "description": "Job URL not found",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    },
                    "500": {
                        "description": "Could not delete Job URL",
                        "schema": {
                            "$ref": "#/definitions/exception.CommonException"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "exception.CommonException": {
            "type": "object",
            "properties": {
                "error_details": {
                    "description": "Optional details about the error\n@example \"User ID does not exist\"",
                    "type": "string"
                },
                "message": {
                    "description": "Error message\n@example \"Resource not found\"",
                    "type": "string"
                },
                "status": {
                    "description": "HTTP status code\n@example 404",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "The time the error occurred\n@example \"2023-10-09T14:48:00Z\"",
                    "type": "string"
                }
            }
        },
...
        "joburl.JobUrl": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "priority": {
                    "description": "Default priority set to 'low'",
                    "type": "string"
                },
                "status": {
                    "description": "Status field, e.g., \"pending\", \"completed\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "description": "URL field with a max length of 2048 characters",
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}